Fix & Refactor: Fix Bug resulting from wrong type and Replace all "match"
Fix a bug resulting from changes in type annotation in
UnsupportedKeyType class in evaluator.
The new annotation for the class expects the argument `type` to be of the type Object. However
it was passed str as it was the old behaviour.

Also refactored evaluator, parser and tokenizer to replace all `match`
statements with if-else ladders for Python 3.7 compatibility