Fix: Unknown AST node error
During the effort to add annotations, some function calls were changed
which resulted in the If-Else node's consequence field to be assigned a
Statement node with its literal value being "if" instead of "" which
resulted in the evaluator trying to execute the Statement node.

This is what caused the error.