fix(mbuiltins, README): Changed the Behaviour of str && Updated README
Previously calling str on a string object just added more double quotes around
the ones which were also there, as guessed, this not the intended
behaviour. Now all double quotes are removed when making a new str using
the builtin str constructor.

Updated the README to add anotheer feature to the Features to be Added,
being optional parametres, it was not there before as I had forgotten
abount it.